+++ src/app/cli/src/cli_parser.rs
@@ -1185,14 +1185,7 @@ pub fn cli() -> Command {
                                     .value_parser(["shell", "json", "yaml"]),
                             ]),
                         Command::new("diagnose")
-                            .about("Run basic system diagnose check")
-                            .args([
-                                Arg::new("output-format")
-                                    .long("output-format")
-                                    .short('o')
-                                    .value_name("FMT")
-                                    .value_parser(["shell", "json", "yaml"]),
-                            ]),
+                            .about("Run basic system diagnose check"),
                         Command::new("ipfs")
                             .about("IPFS helpers")
                             .subcommand_required(true)
