+++ src/app/cli/src/output/mod.rs
@@ -8,8 +8,8 @@
 // by the Apache License, Version 2.0.
 
 mod output_config;
-mod output_write;
+mod verify_progress;
 pub use output_config::*;
-pub use output_write::*;
+pub use verify_progress::*;
 
 pub mod records_writers;
